% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customdists-package.R
\docType{package}
\name{customdists-package}
\alias{customdists}
\alias{customdists-package}
\title{customdists: Custom C++ samplers for various distributions that are missing from the C++ standard library.}
\description{
Custom C++ samplers for various distributions that are missing from the C++ standard library (and boost in many cases)
}
\author{
\strong{Maintainer}: Drew Herren \email{drewherrenopensource@gmail.com} (\href{https://orcid.org/0000-0003-4109-6611}{ORCID})

}
