% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gig.R
\name{sample_gig}
\alias{sample_gig}
\title{Take n draws from a Generalized Inverse Gaussian (GIG)
distribution, using the default parameterization in
https://en.wikipedia.org/wiki/Generalized_inverse_Gaussian_distribution}
\usage{
sample_gig(n, p, a, b)
}
\arguments{
\item{n}{Number of samples to draw}

\item{p}{First parameter of the distribution, which reduces to a shape parameter
in the degenerate case where either of \code{a} or \code{b} is 0 (in which case this
distribution collapses to inverse gamma and gamma, respectively). Can be any real
number but if \code{p = 0} then both \code{a > 0} and \code{b > 0} is required.}

\item{a}{Second parameter of the distribution. Must be nonnegative and cannot be zero
if either of \code{p} or \code{b} is zero.}

\item{b}{Third parameter of the distribution. Must be nonnegative and cannot be zero
if either of \code{p} or \code{a} is zero.}
}
\value{
Vector of length \code{n}
}
\description{
Take n draws from a Generalized Inverse Gaussian (GIG)
distribution, using the default parameterization in
https://en.wikipedia.org/wiki/Generalized_inverse_Gaussian_distribution
}
\examples{
n <- 100
p <- 1.5
a <- 3.4
b <- 0.6
gig_draws <- sample_gig(n,p,a,b)
hist(gig_draws)
}
